+++ source/standalone_examples/api/omni.isaac.quadruped/a1_direct_ros1_standalone.py
@@ -34,12 +34,15 @@ import carb
 
 # enable ROS bridge extension
 enable_extension("omni.isaac.ros_bridge")
+
+simulation_app.update()
+
 # check if rosmaster node is running
 # this is to prevent this sample from waiting indefinetly if roscore is not running
 # can be removed in regular usage
-simulation_app.update()
-result, check = omni.kit.commands.execute("RosBridgeRosMasterCheck")
-if not check:
+import rosgraph
+
+if not rosgraph.is_master_online():
     carb.log_error("Please run roscore before executing this script")
     simulation_app.close()
     exit()
@@ -272,7 +275,7 @@ def main():
     
     """
     # first enable ros node, make sure using simulation time
-    rospy.init_node("isaac_a1", anonymous=False)
+    rospy.init_node("isaac_a1", anonymous=False, disable_signals=True, log_level=rospy.ERROR)
     rospy.set_param("use_sim_time", True)
     physics_downtime = 1 / 400.0
     runner = A1_direct_runner(physics_dt=physics_downtime, render_dt=physics_downtime)