+++ source/standalone_examples/api/omni.isaac.quadruped/go1_ros1_standalone.py
@@ -28,12 +28,15 @@ import carb
 
 # enable ROS bridge extension
 enable_extension("omni.isaac.ros_bridge")
+
+simulation_app.update()
+
 # check if rosmaster node is running
 # this is to prevent this sample from waiting indefinetly if roscore is not running
 # can be removed in regular usage
-simulation_app.update()
-result, check = omni.kit.commands.execute("RosBridgeRosMasterCheck")
-if not check:
+import rosgraph
+
+if not rosgraph.is_master_online():
     carb.log_error("Please run roscore before executing this script")
     simulation_app.close()
     exit()