+++ source/standalone_examples/api/omni.isaac.ros_bridge/clock.py
@@ -17,14 +17,19 @@ import omni
 from omni.isaac.core.utils.extensions import enable_extension
 from omni.isaac.core import SimulationContext
 
+import omni.graph.core as og
+
 # enable ROS bridge extension
 enable_extension("omni.isaac.ros_bridge")
+
+simulation_app.update()
+
 # check if rosmaster node is running
 # this is to prevent this sample from waiting indefinetly if roscore is not running
 # can be removed in regular usage
-simulation_app.update()
-result, check = omni.kit.commands.execute("RosBridgeRosMasterCheck")
-if not check:
+import rosgraph
+
+if not rosgraph.is_master_online():
     carb.log_error("Please run roscore before executing this script")
     simulation_app.close()
     exit()
@@ -32,20 +37,51 @@ if not check:
 from rosgraph_msgs.msg import Clock
 import rospy
 
-# create a clock using sim time
-result, prim = omni.kit.commands.execute(
-    "ROSBridgeCreateClock", path="/ROS_Clock_Sim", clock_topic="/sim_time", sim_time=True
-)
-# create a clock using system time
-result, prim = omni.kit.commands.execute(
-    "ROSBridgeCreateClock", path="/ROS_Clock_System", clock_topic="/system_time", sim_time=False
-)
-# create a clock which we will publish manually, set enabled to false to make it manually controlled
-result, prim = omni.kit.commands.execute(
-    "ROSBridgeCreateClock", path="/ROS_Clock_Manual", clock_topic="/manual_time", sim_time=True, enabled=False
-)
+clock_topic = "sim_time"
+system_clock_topic = "system_time"
+manual_clock_topic = "manual_time"
+
+# Creating a action graph with ROS component nodes
+try:
+    og.Controller.edit(
+        {"graph_path": "/ActionGraph", "evaluator_name": "execution"},
+        {
+            og.Controller.Keys.CREATE_NODES: [
+                ("ReadSimTime", "omni.isaac.core_nodes.IsaacReadSimulationTime"),
+                ("ReadSystemTime", "omni.isaac.core_nodes.IsaacReadSystemTime"),
+                ("OnPlaybackTick", "omni.graph.action.OnPlaybackTick"),
+                ("PublishClock", "omni.isaac.ros_bridge.ROS1PublishClock"),
+                ("PublishSystemClock", "omni.isaac.ros_bridge.ROS1PublishClock"),
+                ("OnImpulseEvent", "omni.graph.action.OnImpulseEvent"),
+                ("PublishManualClock", "omni.isaac.ros_bridge.ROS1PublishClock"),
+            ],
+            og.Controller.Keys.CONNECT: [
+                # Connecting execution of OnPlaybackTick node to PublishClock and PublishSystemClock to automatically publish each frame
+                ("OnPlaybackTick.outputs:tick", "PublishClock.inputs:execIn"),
+                ("OnPlaybackTick.outputs:tick", "PublishSystemClock.inputs:execIn"),
+                # Connecting execution of OnImpulseEvent node to PublishManualClock so it will only publish when an impulse event is triggered
+                ("OnImpulseEvent.outputs:execOut", "PublishManualClock.inputs:execIn"),
+                # Connecting simulationTime data of ReadSimTime to the clock publisher nodes
+                ("ReadSimTime.outputs:simulationTime", "PublishClock.inputs:timeStamp"),
+                ("ReadSimTime.outputs:simulationTime", "PublishManualClock.inputs:timeStamp"),
+                # Connecting systemTime data of ReadSystemTime to the system clock publisher node
+                ("ReadSystemTime.outputs:systemTime", "PublishSystemClock.inputs:timeStamp"),
+            ],
+            og.Controller.Keys.SET_VALUES: [
+                # Assigning topic names to clock publishers
+                ("PublishClock.inputs:topicName", clock_topic),
+                ("PublishSystemClock.inputs:topicName", system_clock_topic),
+                ("PublishManualClock.inputs:topicName", manual_clock_topic),
+            ],
+        },
+    )
+except Exception as e:
+    print(e)
+
+
 simulation_app.update()
 simulation_app.update()
+
 
 # Define ROS callbacks
 def sim_clock_callback(data):
@@ -61,23 +97,24 @@ def manual_clock_callback(data):
 
 
 # Create rospy ndoe
-rospy.init_node("isaac_sim_test_gripper", anonymous=True, disable_signals=True, log_level=rospy.ERROR)
+rospy.init_node("isaac_sim_clock", anonymous=True, disable_signals=True, log_level=rospy.ERROR)
 # create subscribers
-sim_clock_sub = rospy.Subscriber("sim_time", Clock, sim_clock_callback)
-system_clock_sub = rospy.Subscriber("system_time", Clock, system_clock_callback)
-manual_clock_sub = rospy.Subscriber("manual_time", Clock, manual_clock_callback)
+sim_clock_sub = rospy.Subscriber(clock_topic, Clock, sim_clock_callback)
+system_clock_sub = rospy.Subscriber(system_clock_topic, Clock, system_clock_callback)
+manual_clock_sub = rospy.Subscriber(manual_clock_topic, Clock, manual_clock_callback)
 time.sleep(1.0)
 simulation_context = SimulationContext(physics_dt=1.0 / 60.0, rendering_dt=1.0 / 60.0, stage_units_in_meters=1.0)
 # need to initialize physics getting any articulation..etc
 simulation_context.initialize_physics()
+
 simulation_context.play()
 
 # perform a fixed number of steps with fixed step size
 for frame in range(20):
 
     # publish manual clock every 10 frames
-    if frame % 2 == 0:
-        result, status = omni.kit.commands.execute("RosBridgeTickComponent", path="/ROS_Clock_Manual")
+    if frame % 10 == 0:
+        og.Controller.set(og.Controller.attribute("/ActionGraph/OnImpulseEvent.state:enableImpulse"), True)
         simulation_context.render()  # This updates rendering/app loop which calls the system and sim clocks
 
     simulation_context.step(render=False)  # runs with a non-realtime clock
@@ -89,8 +126,8 @@ for frame in range(20):
 for frame in range(20):
 
     # publish manual clock every 10 frames
-    if frame % 2 == 0:
-        result, status = omni.kit.commands.execute("RosBridgeTickComponent", path="/ROS_Clock_Manual")
+    if frame % 10 == 0:
+        og.Controller.set(og.Controller.attribute("/ActionGraph/OnImpulseEvent.state:enableImpulse"), True)
 
     simulation_app.update()  # runs with a realtime clock
     # This sleep is to make this sample run a bit more deterministically for the subscriber callback